ml/requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow

supabase>=2.0

scikit-learn>=1.3
scipy>=1.11
implicit

sentence-transformers>=2.7
torch

faiss-cpu

fastapi>=0.100
uvicorn>=0.23

requests
joblib
python-dotenv
tqdm